moving reservation buttons to be a part of calendar and adding hook to refresh calendar when a booking is made
