+++ src/components/Bookings/Calendar.tsx
@@ -1,13 +1,15 @@
-import React, { useState, useEffect } from 'react';
+import React, { useState, useEffect, useRef } from 'react';
+import '../../stylings/ReservationButtons.css'
 import { Calendar as BigCalendar, momentLocalizer, Event, Views } from 'react-big-calendar';
 import moment from 'moment';
 import 'react-big-calendar/lib/css/react-big-calendar.css';
-import { collection, getDocs, DocumentData } from 'firebase/firestore';
+import { collection, getDocs, Timestamp, addDoc } from 'firebase/firestore';
 import { db } from '../../firebase';
-import ReservationButtons from './ReservationButtons';
 
 const localizer = momentLocalizer(moment);
 const ref = collection(db, "bookings");
+
+
 
 export default function Calendar() {
 
@@ -28,9 +30,34 @@ export default function Calendar() {
     setLoading(false); // Indicate loading is done
   };
 
-useEffect(() => {
-  getBookings(); // Fetch bookings when the component mounts
-}, []); 
+  useEffect(() => {
+    getBookings(); // Fetch bookings when the component mounts
+  }, []);
+
+  const startTime = useRef<HTMLInputElement>(null);
+  const endTime = useRef<HTMLInputElement>(null);
+
+  const handleBooking = async (e: React.FormEvent) => {
+    e.preventDefault();
+    
+    if (startTime.current && endTime.current) {
+      const start = new Date(startTime.current.value);
+      const end = new Date(endTime.current.value);
+      let data = {
+          startTime: Timestamp.fromDate(start),
+          endTime: Timestamp.fromDate(end)
+      };
+  
+        try {
+            await addDoc(ref, data);
+            await getBookings();
+            console.log("Booking added!");
+        } catch (error) {
+            console.error("Error adding booking: ", error);
+        }
+    }
+  };
+
   return (
 
     <div
@@ -53,7 +80,18 @@ useEffect(() => {
           style={{ height: '100%', width: '100%' }} 
         />
       ) }
-      <ReservationButtons />
+
+      <div style={{ display: 'flex', justifyContent: 'space-between', width: '50%', margin: '20px auto' }}>
+        <input type="datetime-local" ref={startTime} placeholder="Enter the start time" required />
+        <input type="datetime-local" ref={endTime} placeholder="Enter the end time" required />
+      </div>
+
+      {/* Buttons */}
+      <div style={{ display: 'flex', justifyContent: 'space-between', width: '50%', margin: '20px auto' }}>
+        <button className="button" onClick={handleBooking}>Make a Reservation</button>
+        <button className="button" onClick={() => console.log('Delete clicked')}>Cancel Reservation</button>
+        <button className="button" onClick={() => console.log('Update clicked')}>Modify Reservation</button>
+      </div>
     </div>
   );
 }
